mypy modifications

+++ backend/crud/user.py
@@ -1,17 +1,14 @@
 # crud.py
 from sqlalchemy.orm import Session
-from models import User
-from schemas import UserCreate, UserUpdate
-from models import pwd_context
-
+from models.user import User
+from schemas.user import UserCreate, UserUpdate
+from models.user import pwd_context
 
 def get_user(db: Session, user_id: int):
     return db.query(User).filter(User.id == user_id).first()
 
-
 def get_users(db: Session, skip: int = 0, limit: int = 10):
     return db.query(User).offset(skip).limit(limit).all()
-
 
 def create_user(db: Session, user: UserCreate):
     hashed_password = pwd_context.hash(user.password)
@@ -20,30 +17,33 @@ def create_user(db: Session, user: UserCreate):
         email=user.email,
         full_name=user.full_name,
         password_hashed=hashed_password,
+        adresse=user.adresse,
+        phone=user.phone,
     )
     db.add(db_user)
     db.commit()
     db.refresh(db_user)
     return db_user
 
-
-def update_user(db: Session, user_id: int, user_update: UserUpdate):
+def update_user(db: Session, user_id: int, user_update: dict):
     db_user = db.query(User).filter(User.id == user_id).first()
-    if user_update.password:
-        hashed_password = pwd_context.hash(user_update.password)
-        db_user.password_hashed = hashed_password
-    if user_update.username:
-        db_user.username = user_update.username
-    if user_update.email:
-        db_user.email = user_update.email
-    if user_update.full_name:
-        db_user.full_name = user_update.full_name
+    if db_user is None:
+        raise ValueError("User not found")
+
+    for key, value in user_update.items():
+        if key == 'password':
+            hashed_password = pwd_context.hash(value)
+            setattr(db_user, 'password_hashed', hashed_password)
+        else:
+            setattr(db_user, key, value)
+
     db.commit()
     db.refresh(db_user)
     return db_user
 
-
 def delete_user(db: Session, user_id: int):
     db_user = db.query(User).filter(User.id == user_id).first()
+    if db_user is None:
+        raise ValueError("User not found")
     db.delete(db_user)
     db.commit()
